# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "AI Train VM API.",
        "title": "rfstack API",
        "contact": {},
        "license": {},
        "version": "0.2"
    },
    "host": "localhost:8088",
    "basePath": "/",
    "paths": {
...
        "/v1/flavor/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List flavor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Flavor"
                ],
                "summary": "List flavor",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.FlavorsListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/image/list": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "List image",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Image"
                ],
                "summary": "List image",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.ImagesListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/docs.GenericErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "docs.FlavorsListResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "flavors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.LabelValue"
                    }
                }
            }
        },
        "docs.GenericErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean",
                    "format": "bool",
                    "example": true
                },
                "message": {
                    "type": "string",
                    "format": "string",
                    "example": "error response message"
                }
            }
        },
...
        "docs.ImagesListResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "boolean"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/docs.LabelValue"
                    }
                }
            }
        },
...
        "docs.LabelValue": {
            "type": "object",
            "properties": {
                "label": {
                    "type": "string",
                    "example": "name"
                },
                "value": {
                    "type": "string",
                    "example": "32482124-6d7d-47a8-b4a9-dea50e50823f"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}